Fixed bad EUCovid test structure
