+++ ts/features/euCovidCert/__e2e__/euCovidCertExpired00.e2e.ts
@@ -1,9 +1,11 @@
 import { device } from "detox";
 import { ensureLoggedIn } from "../../../__e2e__/utils";
+import { e2eWaitRenderTimeout } from "../../../__e2e__/config";
+import { learnMoreLinkTestId, scrollToEUCovidMessage } from "./utils";
 
 const euCovidCertExpiredSubject = "🏥 EUCovidCert - expired";
-
-const messageListTestId = "MessageList_inbox";
+const euCovidCertExpiredTitle = "Expired Certificate title";
+const euCovidCertExpiredSubTitle = "Expired Certificate sub title";
 
 describe("EuCovidCert Expired", () => {
   beforeAll(async () => {
@@ -11,13 +13,26 @@ describe("EuCovidCert Expired", () => {
     await ensureLoggedIn();
   });
 
-  it("should find the expired EuCovidCert message and open it", async () => {
-    await waitFor(element(by.text(euCovidCertExpiredSubject)))
-      .toBeVisible()
-      .whileElement(by.id(messageListTestId))
-      .scroll(350, "down");
+  it("should find the expired EuCovidCert message, open it and check all the correct elements in the details page", async () => {
+    await openExpiredEUCovidMessage();
 
-    const subject = element(by.text(euCovidCertExpiredSubject));
-    await subject.tap();
+    await waitFor(element(by.text(euCovidCertExpiredTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.text(euCovidCertExpiredSubTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.id(learnMoreLinkTestId)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
   });
 });
+
+const openExpiredEUCovidMessage = async () => {
+  await scrollToEUCovidMessage(euCovidCertExpiredSubject);
+
+  const subject = element(by.text(euCovidCertExpiredSubject));
+  await subject.tap();
+};
