+++ ts/features/euCovidCert/__e2e__/euCovidCertRevoked00.e2e.ts
@@ -1,9 +1,11 @@
 import { device } from "detox";
 import { ensureLoggedIn } from "../../../__e2e__/utils";
+import { e2eWaitRenderTimeout } from "../../../__e2e__/config";
+import { learnMoreLinkTestId, scrollToEUCovidMessage } from "./utils";
 
 const euCovidCertRevokedSubject = "🏥 EUCovidCert - revoked";
-
-const messageListTestId = "MessageList_inbox";
+const euCovidCertRevokedTitle = "Revoked Certificate title";
+const euCovidCertRevokedSubTitle = "Revoked Certificate sub title";
 
 describe("EuCovidCert Revoked", () => {
   beforeAll(async () => {
@@ -11,13 +13,26 @@ describe("EuCovidCert Revoked", () => {
     await ensureLoggedIn();
   });
 
-  it("should find the revoked EuCovidCert message and open it", async () => {
-    await waitFor(element(by.text(euCovidCertRevokedSubject)))
-      .toBeVisible()
-      .whileElement(by.id(messageListTestId))
-      .scroll(350, "down");
+  it("should find the revoked EuCovidCert message, open it and check all the correct elements in the details page", async () => {
+    await openRevokedEUCovidMessage();
 
-    const subject = element(by.text(euCovidCertRevokedSubject));
-    await subject.tap();
+    await waitFor(element(by.text(euCovidCertRevokedTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.text(euCovidCertRevokedSubTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.id(learnMoreLinkTestId)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
   });
 });
+
+const openRevokedEUCovidMessage = async () => {
+  await scrollToEUCovidMessage(euCovidCertRevokedSubject);
+
+  const subject = element(by.text(euCovidCertRevokedSubject));
+  await subject.tap();
+};
