+++ ts/features/euCovidCert/__e2e__/euCovidCertValid00.e2e.ts
@@ -1,9 +1,11 @@
 import { device } from "detox";
+import I18n from "../../../i18n";
 import { ensureLoggedIn } from "../../../__e2e__/utils";
+import { e2eWaitRenderTimeout } from "../../../__e2e__/config";
+import { openValidEUCovidMessage, qrCodeTestId } from "./utils";
 
-const euCovidCertValidSubject = "🏥 EUCovidCert - valid";
-
-const messageListTestId = "MessageList_inbox";
+const euCovidCertValidTitle = "Valid Certificate title";
+const euCovidCertValidSubTitle = "Valid Certificate sub title";
 
 describe("EuCovidCert Valid", () => {
   beforeAll(async () => {
@@ -11,13 +13,23 @@ describe("EuCovidCert Valid", () => {
     await ensureLoggedIn();
   });
 
-  it("should find the valid EuCovidCert message and open it", async () => {
-    await waitFor(element(by.text(euCovidCertValidSubject)))
-      .toBeVisible()
-      .whileElement(by.id(messageListTestId))
-      .scroll(350, "down");
+  it("should find the valid EuCovidCert message, open it and check all the correct elements in the details page", async () => {
+    await openValidEUCovidMessage();
 
-    const subject = element(by.text(euCovidCertValidSubject));
-    await subject.tap();
+    await waitFor(element(by.text(euCovidCertValidTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.text(euCovidCertValidSubTitle)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.text(I18n.t("global.genericSave"))))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
+
+    await waitFor(element(by.id(qrCodeTestId)))
+      .toBeVisible()
+      .withTimeout(e2eWaitRenderTimeout);
   });
 });
