+++ ts/features/euCovidCert/__e2e__/euCovidCertValid01.e2e.ts
@@ -2,11 +2,11 @@ import { device } from "detox";
 import I18n from "../../../i18n";
 import { e2eWaitRenderTimeout } from "../../../__e2e__/config";
 import { ensureLoggedIn } from "../../../__e2e__/utils";
-
-const euCovidCertValidTitle = "Valid Certificate title";
-const euCovidCertValidSubTitle = "Valid Certificate sub title";
-
-const qrCodeTestId = "QRCode";
+import {
+  fullScreenQrCodeTestId,
+  openValidEUCovidMessage,
+  qrCodeTestId
+} from "./utils";
 
 describe("EuCovidCert Valid", () => {
   beforeAll(async () => {
@@ -14,21 +14,21 @@ describe("EuCovidCert Valid", () => {
     await ensureLoggedIn();
   });
 
-  it("should check all the correct elements in the details page", async () => {
-    await waitFor(element(by.text(euCovidCertValidTitle)))
+  it("should open the QRCode in fullscreen and return back", async () => {
+    await openValidEUCovidMessage();
+
+    const qrCode = element(by.id(qrCodeTestId));
+    await qrCode.tap();
+
+    await waitFor(element(by.id(fullScreenQrCodeTestId)))
       .toBeVisible()
       .withTimeout(e2eWaitRenderTimeout);
 
-    await waitFor(element(by.text(euCovidCertValidSubTitle)))
+    await waitFor(element(by.text(I18n.t("global.buttons.close"))))
       .toBeVisible()
       .withTimeout(e2eWaitRenderTimeout);
 
-    await waitFor(element(by.text(I18n.t("global.genericSave"))))
-      .toBeVisible()
-      .withTimeout(e2eWaitRenderTimeout);
-
-    await waitFor(element(by.id(qrCodeTestId)))
-      .toBeVisible()
-      .withTimeout(e2eWaitRenderTimeout);
+    const closeButton = element(by.text(I18n.t("global.buttons.close")));
+    await closeButton.tap();
   });
 });
